fixup! EZP-29998: As a developer, I want to define icons for content types
