+++ src/lib/UI/Config/Provider/ContentTypes.php
@@ -9,6 +9,7 @@ namespace EzSystems\EzPlatformAdminUi\UI\Config\Provider;
 use eZ\Publish\API\Repository\ContentTypeService;
 use eZ\Publish\Core\MVC\Symfony\Locale\UserLanguagePreferenceProviderInterface;
 use EzSystems\EzPlatformAdminUi\UI\Config\ProviderInterface;
+use EzSystems\EzPlatformAdminUi\UI\Service\ContentTypeIconResolver;
 
 class ContentTypes implements ProviderInterface
 {
@@ -18,14 +19,22 @@ class ContentTypes implements ProviderInterface
     /** @var \eZ\Publish\Core\MVC\Symfony\Locale\UserLanguagePreferenceProviderInterface */
     private $userLanguagePreferenceProvider;
 
+    /** @var \EzSystems\EzPlatformAdminUi\UI\Service\ContentTypeIconResolver */
+    private $contentTypeIconResolver;
+
     /**
      * @param \eZ\Publish\API\Repository\ContentTypeService $contentTypeService
      * @param \eZ\Publish\Core\MVC\Symfony\Locale\UserLanguagePreferenceProviderInterface $userLanguagePreferenceProvider
+     * @param \EzSystems\EzPlatformAdminUi\UI\Service\ContentTypeIconResolver $contentTypeIconResolver
      */
-    public function __construct(ContentTypeService $contentTypeService, UserLanguagePreferenceProviderInterface $userLanguagePreferenceProvider)
-    {
+    public function __construct(
+        ContentTypeService $contentTypeService,
+        UserLanguagePreferenceProviderInterface $userLanguagePreferenceProvider,
+        ContentTypeIconResolver $contentTypeIconResolver
+    ) {
         $this->contentTypeService = $contentTypeService;
         $this->userLanguagePreferenceProvider = $userLanguagePreferenceProvider;
+        $this->contentTypeIconResolver = $contentTypeIconResolver;
     }
 
     /**
@@ -48,6 +57,7 @@ class ContentTypes implements ProviderInterface
                 $contentTypeGroups[$contentTypeGroup->identifier][] = [
                     'identifier' => $contentType->identifier,
                     'name' => $contentType->getName(),
+                    'thumbnail' => $this->contentTypeIconResolver->getContentTypeIcon($contentType),
                 ];
             }
         }
